two methods moved from validation_run_view.py to modify_validation_view.py as they fit there better
